error handlerの位置を移動

+++ src/application/service/user.ts
@@ -2,6 +2,7 @@ import {User} from '@/domain/entity/index'
 import {UserRepository, RoleRepository} from '@/domain/types/repository'
 import {RoleView, UsersView, UserView} from '@/application/types/view'
 import {UserPresenter, RolePresenter} from '@/application/presenter/index'
+import {ErrorBadRequest} from '@/domain/entity/index'
 
 export default class UserService {
   readonly userRepository: UserRepository
@@ -40,7 +41,11 @@ export default class UserService {
     roleIDs: string[]
   ): Promise<UserView> {
     const user = new User(new Date(), email, password, name, roleIDs)
+
+    const exist = await this.userRepository.findById(user.id)
+    if (exist) throw new ErrorBadRequest(409, 'already exists')
     await this.userRepository.create(user)
+
     const rolePresenter = new RolePresenter()
     const roles = await this.roleRepository.findAll()
         .then(roles => roles.map(role => {
@@ -59,6 +64,7 @@ export default class UserService {
 
   async getUser(id: string) {
     const user = await this.userRepository.findById(id)
+    if (!user) throw new ErrorBadRequest(404, 'resource not found')
     const presenter = await this.userPresenter()
     const res = presenter.resolve(user)
     return res
@@ -72,6 +78,7 @@ export default class UserService {
     roleIDs: string[]
   ) {
     const user = await this.userRepository.findById(id)
+    if (!user) throw new ErrorBadRequest(404, 'resource not found')
     user.update(email, password, name, roleIDs)
     await this.userRepository.update(user)
     const presenter = await this.userPresenter()
@@ -80,6 +87,8 @@ export default class UserService {
   }
 
   async deleteUser(id: string) {
+    const user = await this.userRepository.findById(id)
+    if (!user) throw new ErrorBadRequest(404, 'resource not found')
     await this.userRepository.delete(id)
   }
 }
